openstack: Add context for SRIOV configuration
Implement the ``neutron-openvswitch`` charm ``configure_sriov``
helper function as a context.

This is useful both for the interrim and long term use case where
we now will write a interface name and number of VFs to a YAML
file.

As of this writing it will be used with the ``sriov-netplan-shim``
package, long term it can be re-used to augment Netplan config.

Co-Authored-By: James Page <james.page@ubuntu.com>