ceph autoscaler: set min pg num for pools
Ensure that the minimum number of pgs is set as appropriate during
pool creation. If the number of pgs used for the initial pool creation
is less that the autoscaler default of 32, set to the lower pg
number to ensure that the autoscaler does not immediately scale
the pool up to 32 pgs.

This avoids pg overloading in smaller test deployments.

Closes-Bug: 1872748