Add support for Erasure Coded pools
Add support in the ceph module for creation of new erasure coding
profiles and support for setting the allow_ec_overwrites property
on erasure coded pools.  This is required to support RBD use
cases.

Update OpenStack Ceph context to detect use of EC pool type,
providing an appropriate value for the rbd default data pool
for use in ceph.conf.

This last part is a bit of a workaround due to the fact that
OpenStack services don't have native configuration options to
set the metadata and data pool options when creating new RBD
images.

Add support for configuration and validation of the erasure
plugin technique as part of the erasure profile creation.

Add support for extended configuration of specific erasure code
plugins.