Gate on 'password' from rabbit units for amqp ctxt (#482)
Currently a rabbit host is added to the list of available hosts
in the rabbit context if the `private-address` key is available on
the relation. But at the start of the relation `private-address` is
always present, even before the corresponding `amqp-relation-joined`
hook has run on the rabbit unit. So the rabbit unit has no control
over when that is first presented as its set by juju itself (the
value of `private-address` is subsequently overridden by the
charm but the charm does not control when it first appears).

This change checks that the rabbit unit has set `password` on
the relation before adding it to the list of available rabbit
units. The setting of `password` is entirely within the control
of the rabbit charm and the rabbit charm should ensure it is only
set when the corresponding unit is ready and clustered.