core.hookenv.Config: Handle invalid state files (#484)
When initialising core.hookenv.Config a blank or corrupt state file
(.juju-persistent-config) caused the following exception to be raised:

  File "charmhelpers/core/hookenv.py", line 348, in load_previous
    for k, v in copy.deepcopy(self._prev_dict).items():
AttributeError: 'NoneType' object has no attribute 'items'

Handle this situation and return a blank previous state instead to
recover from this situation.

A previous attempt to fix this was made in #166 (LP #1768018) however it
did not work because while it caught the error loading the JSON file, it
then tried to call .items() on the object copy which is set to None.
Fix that by returning from the function after catching the exception.

The test cases (test_init_empty_state_file,
test_init_invalid_state_file) were also incorrectly passing because they
wrote to the test state file after initialising the Config object when
it had already been read. Fix that by writing to the file before Config
is initialised instead.

Note that in such a case where the file was corrupt, we will log an
error but otherwise silently "fix" this by loading a blank dictionary
causing Config() to report that all config values have changed even
though they may not have. Even though the code already intended (but
failed) to do this, we should ensure that is acceptable when reviewing
this change. I have updated the logged error to make this explicit.