contrib/ovs: Fix deletion of virtual bridge ports (#491)
The OVS `del_bridge_port` helper currently assumes it will be used
with interfaces that actually exist in the system.

This is not always the case. Add arguments that allow the caller
to decide whether a interface link and promiscuous settings should
be touched after removing the port from a OVS bridge. Defaults to
being compatible with the current behaviour.