contrib/ceph: Share common post pool creation code
Deprecate the current ``Pool`` class and replace it with a
``BasePool`` class to keep compatibility with any external
consumers.

The new class moves the create and post-create steps into the
private ``_create`` and ``_post_create`` methods that the pool
type specific children override.

This allows for common post-creation steps to be shared.