Correct name of CA cert file
The CA file received down the certificates relation is currently
named keystone_juju_ca_cert.pem. The is left over from the old
cert system where keystone would act as a CA. Since the CA has
nothing to do with keystone, rename the cert file to include
the name of the remote application which provided the certificate.