Fix race in checking closed ports on pause
When pausing a service a check is called to verify if services are
stopped and ports are closed. This is prone to racing as service
shutdown is not guaranteed to be synchronous.

This change implements a helper to wait for ports to be closed which
clients can use to mitigate this race.