Add max_check_attempts support for nrpe checks (#539)
Add support to nrpe for optional attribute max_check_attempts.  If
supplied, max_check_attempts is added to the NRPE check when consumed by
the Nagios host, overriding the default.

contrib.charmsupport.nrpe writes service definition files for Nagios to
collect.  This adds max_check_attempts to that file should it be
specified.