Add support to cert_utils for bespoke bindings
The cert_utils use ADDRESS_MAP as a list a definitive list of
bindings in various places. This PR allows a charm to extend that
map with its own bindings. It also allows the 'config' and 'override' keys to
be absent. For some bindings there is no corresponding charm config options
so it is reasonable for them to be absent.

* Update get_certificate_request, get_certificate_sans and create_ip_cert_links
  to accept an address_map kwarg to allow bespoke bindings to be resolved.
* Fix bug in get_certificate_request which removes an IP from _sans even
  if it has not yet been added to the certificate request (occurs if an ip
  is in _sans but if hostname_override is None)
* Update _get_address_override to accept an address_map kwarg which can then
  be used to search for override key.
* Update _get_address_override to allow the override key to be missing and
  return None if it is.
* Update resolve_address to accept an address_map kwarg.
* Update resolve_address to allow the 'config' key to be missing from the
  address map.