Cap the number of workers always when multiplier is not set
Previously the cap is only applied to the units in containers. However,
services in a bare metal also requires a sensible cap. Otherwise,
nova-compute, for example, will have 256 workers for nova-api-metadata
out of the box which is an overkill with the following system.

32 cores * 2 threads/core * 2 sockets * 2 (default multiplier) = 256

Let's cap the number of workers as 16 for bare metal and 4 for
containers, then let operators override it with an explicit config to
worker-multiplier.

Closes-Bug: LP: #1843011