Update ubuntu_apt_pkg.py (#559)
The first part of dpkg-query's output is the "Desired" status of the package which can be one of "Unknown/Install/Remove/Purge/Hold". Where "Hold" means "A package marked to be on hold is kept on the same version" which is a variation of installed. Subsequently, a package with status of Hold should be considered as installed.

Co-authored-by: David Ames <david.ames@canonical.com>