Use HA peers as memcache cluster
Charms using `MemcacheContext`deploy local `memcached` services on
each unit but are not utilizing a `oslo_cache.memcache_pool`. This can
lead to inconsistently cached values and consequently to incorrect
query results.

This change add the `memcache` relation to connect to the memcached
charm and uses the address of the memcached units instead.

Closes-Bug: #1771114

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>