Switch charm events to use policy_rcd file backend
The previous implementation used two different backends for storing
deferred events. If the event came from a charm it would be stored
in the charms sqlite db. If the event came from a package it would
be stored as file. To simplify things this patch switched the charm
to use the same file backend as the packaging.

This change has involved some refactoring and the movement of the
functions around mamanging of events from policy_rcd.py to
deferred_events.py since the events are no longer specific to
policy_rcd.