Adds context switch for neutron-dhcp-agent
The neutron-dhcp-agent might rely on options defined in
openvswitch_agent.ini. By default this config file is not passed to
neutron-dhcp-agent daemon, and therefore those options are not
loaded and fall back for the default values and by that inhibits
the desired behavior. This behavior should happen only if
the release if greater than queens.

Please check: lp:#1832021 and gerrit:I134c8077ee52ccdb4e383109ecbea27ed1633fb8
for more context.