Add a base openstack code name as a source for add_source()
This is a 'smart' option, where the openstack-origin=ussuri, for
example, passing 'ussuri' to the add_source() function.  This, on focal,
will be the equivalent of 'distro' and on bionic, will be
'cloud:bionic-ussuri'.  On groovy, victoria is distro, etc.

Also support a 'ussuri/proposed' to allow proposed to be carried between
distro versions.  This is to simplify the logic for setting openstack
versions so that the version wanted is the name rather than 'distro'.