ceph: make ceph-common version checks optional
Its possible that the ceph module will be used on units where
ceph-common is not installed in order to create requests to
send to the Ceph Broker (hosted by the ceph-mon charm).

Avoid any hard failures when checking the version and assume
and older version that nautilus is actually installed.