Add force_upgrade parm to openstack-upgrade action (#649)
* Add dist_upgrade parm to openstack-upgrade action

If dist_upgrade is set to true, the dist-upgrade will be performed
regardless of whether a new OpenStack release is available. This
can be useful to dist-upgrade to a proposed pocket or PPA for
testing, or to upgrade packages to a new stable version.

* Add force_upgrade parm to openstack-upgrade action

If force_upgrade is set to true, the upgrade will be performed
regardless of whether a new OpenStack release is available. This
can be useful for upgrading to the latest stable packages, or for
testing upgrades to a proposed pocket or PPA of the same release.