Fix handling of profile-name
The current code:
self.profile_name = op.get('crush-profile', profile_name)

will only default to profile_name if the 'crush-profile' key
doesn't exist in the op dictionary. If the 'crush-profile' key
exists and is set to None, the default profile_name is not used.

This change will use the default profile_name in both cases.

Closes-Bug: #1960622