Modify git tests so they continue to work with git < 2.28
git 2.28 introduced the `--initial-branch` option to
`git init` which allowed the initial branch to be other than
master.  This was used in the GitUrlFetchHandlerTest class.
Unfortunately, this breaks when testing on a distro that doesn't
have git < 2.28 installed.  Modify the tests so that this is called
out explicitly.  When git pre-2.28 is gone, this patch can be
revered.