Don't run 'apt install' when list of packages is empty.
A common pattern in charms is to call `apt_install()` passing a filtered
list of packages to avoid reinstallation of packages, so in normal
circumstances a charm may end up calling apt_install() with an empty
list producing the execution of "apt-get install" which will produce at
the very least a read of the list of packages database and build the
dependency tree, this is something that for IO constrained environments
impacts the performance of the running services. Also for environments
where the package database is broken (e.g. a post install script failed)
this would add more entropy and potentially making things more difficult
for the operator to fix.

$ sudo apt-get install
Reading package lists... Done
Building dependency tree... Done
Reading state information... Done
0 upgraded, 0 newly installed, 0 to remove and 63 not upgraded.