openstack: identity-service forwards compat
The new Keystone K8S operator exclusively uses the Application
data bag to provide endpoint information and credentials to
consuming services.

Check for the presence of data in the App data bag and use this
in preference to per-unit data as presented by the Keystone
machine charm.

This also requires some remapping and manipulation to massage
the new application data bag keys into the existing ctxt dict
used in templates.