Add support for OVN UCA pocket
The OVN UCA pocket will be used as an overlay repository in
addition to any regular UCA pocket and the charms are expected
to expose configuration of it as a separate source configuration
option.

Add pocket names and allow for writing any OVN UCA configuration
to a separate sources.list.d file.