unitdata: make keeping revision data optional (#737)
The unitdata code keeps variable revisions in kv_revisions table, and
as a part of that all of the environment variables serialized into JSON and
pushed into the table if current set of environment variables differ from the
previous ones. But the code considers environment variable list as a whole and
even if a single environment variable changes, whole list gets pushed as a
revision. Some of the environment variables (e.g. JUJU_CONTEXT_ID) is always
changing, which means whole env variable list is serialized and pushed to
the DB on every single hook invocation. This has become a problem in deployments
with very large (e.g. ~70 KiB) environment variable lists. Given that
update-status is running periodically (every 5 mins by default) and
many charms running in the same environment, the disk space of the host
running the charms runs out over time, raising a need of manual intervention
to clear up the kv_revisions table.

This fix makes keeping revisions optional, and disabled by default.

Closes-bug: LP#1930173