Update method for checking endpoint protocol (#769)
Update method for checking endpoint protocol

The `https` method is used to check if an endpoint is expected to
be http or https. One of the checks it performs is to examine the
the certificates relation. If the relation is present then it looks
for the existance of a CA. However the OpenStack charms do not
switch to https until a certificate is provided via the certificates
relation. This means there can be a disconnect if the
certificate provider has provided a CA but has not yet provided
the unit specific certificates. If this happens then the payload
will still be using http but the `https` method will return True.

This patch updates the `https` method to return False if an unfilled
certificate request exists.

(cherry picked from commit 6064a34627882d1c8acf74644c48d05db67ee3b4)