Drop use of get_os_version_codename_swift
Swift payload upgrades to wallaby or later currently fail because
this code is unable to determine the version for wallaby+. For
example:

FATAL ERROR: Could not derive swift version for codename: wallaby

We stopped maintaining the SWIFT_CODENAMES and PACKAGE_CODENAMES as of
wallaby because the openstack-release package was introduced in wallaby.

We could update the SWIFT_CODENAMES map, but really this map no longer
needs to be used. There was a time when the same swift package version
existed in two different OpenStack releases, but that is no longer the
case [1]. Therefore we can use the same comparison used for OpenStack
packages which essentially checks current release (based on the
openstack-release package) vs the installation source release.

[1] Current swift versions in Ubuntu are:
2.32.0 bobcat
2.31.1 antelope
2.30.1 zed
2.29.2 yoga
2.28.1 xena
2.27.0 wallaby
2.26.0 victoria
2.25.2 ussuri

Closes-Bug: #2040606