Fix dynamic ipv6 address check & DNS tests
get_ipv6_addr: the if condition was checking for the eui 64 mac address, but that has nothing to do with global ipv6 addresses.
I added and fixed existing tests accordingly.
The DNS resolver tests were not expecting a NoAnswer that is returned if the query is done against a non-authoritative recursor.
Finally fixed a typo in a test name: loopup -> lookup.