Allow non-dynamic addresses for ipv6 relation_ips
When looking for relation IP addresses, the current code filters
for only dynamic addresses. This doesn't work on MAAS where it
assigns fixed-address6 for the dhcpd6 configuration - e.g.

host 5c-ed-8c-e8-d5-b4 {
   #
   # Node DHCP snippets
   #
   # No DHCP snippets defined for host

   hardware ethernet 5c:ed:8c:e8:d5:b4;
   fixed-address6 fc00:1:1::2;
}

This causes the local node to not have the address as dynamic. Change
the default option to set dynamic_only=False

Co-Authored-by: Matthieu Marchand <matthieu.marchand@canonical.com>
Signed-off-by: Billy Olsen <billy.olsen@canonical.com>